app/requirements.txt
argcomplete==1.10.3
attrs==21.2.0
azure-ai-textanalytics==5.1.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.8.2
black==21.9b0
cfgv==3.3.1
chardet==3.0.4
click==8.0.1
colorama==0.4.4
compressed-rtf==1.0.6
coverage==6.0.1
distlib==0.3.3
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
extract-msg==0.28.7
filelock==3.3.0
flake8==3.9.2
identify==2.3.0
IMAPClient==2.1.0
iniconfig==1.1.1
isort==5.9.3
joblib==1.0.1
lxml==4.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
nltk==3.6.2
nodeenv==1.6.0
olefile==0.46
packaging==21.0
pathspec==0.9.0
pdfminer.six==20191110
Pillow==9.0.0
platformdirs==2.3.0
pluggy==1.0.0
pre-commit==2.15.0
py==1.10.0
pycodestyle==2.7.0
pycryptodome==3.12.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2021.8.28
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
textblob==0.15.3
textract==1.6.4
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
virtualenv==20.8.1
xlrd==1.2.0
XlsxWriter==3.0.2
